+++ app/src/main/java/com/example/intels_app/ProfileAdapterAdmin.java
@@ -8,6 +8,7 @@ package com.example.intels_app;
 
 import static android.content.ContentValues.TAG;
 
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -26,6 +27,7 @@ import com.bumptech.glide.Glide;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.storage.FirebaseStorage;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -69,25 +71,40 @@ public class ProfileAdapterAdmin extends BaseAdapter {
         deleteButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                // Remove item from Firestore, data list, and notify adapter
-
-                FirebaseFirestore.getInstance().collection("profiles").document(profile.getName())
-                        .delete()
-                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                            @Override
-                            public void onSuccess(Void unused) {
-                                Log.d(TAG, "DocumentSnapshot successfully deleted!");
+                new AlertDialog.Builder(context)
+                        .setTitle("Confirm Deletion")
+                        .setMessage("Are you sure you want to delete this profile?")
+                        .setPositiveButton("Yes", (dialog, which) -> {
+                            // Remove item from Firestore, data list, and notify adapter
+                            if (profile.getImageUrl() != null && !profile.getImageUrl().isEmpty()) {
+                                FirebaseStorage.getInstance().getReferenceFromUrl(profile.getImageUrl()).delete()
+                                        .addOnSuccessListener(unused -> Log.d(TAG, "Image successfully deleted."))
+                                        .addOnFailureListener(e -> Log.w(TAG, "Failed to delete image.", e));
                             }
-                        }).addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Log.w(TAG, "Error deleting document", e);
-                            }
-                        });
 
-                profiles.remove(position);
-                notifyDataSetChanged();
-                Toast.makeText(context, "Event deleted", Toast.LENGTH_SHORT).show();
+                            FirebaseFirestore.getInstance().collection("profiles").document(profile.getName())
+                                    .delete()
+                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                        @Override
+                                        public void onSuccess(Void unused) {
+                                            Log.d(TAG, "DocumentSnapshot successfully deleted!");
+                                        }
+                                    }).addOnFailureListener(new OnFailureListener() {
+                                        @Override
+                                        public void onFailure(@NonNull Exception e) {
+                                            Log.w(TAG, "Error deleting document", e);
+                                        }
+                                    });
+
+                            profiles.remove(position);
+                            notifyDataSetChanged();
+                            Toast.makeText(context, "Profile deleted", Toast.LENGTH_SHORT).show();
+                        })
+                        .setNegativeButton("No", (dialog, which) -> {
+                            // Dismiss the dialog if the user cancels
+                            dialog.dismiss();
+                        })
+                        .show();
             }
         });
